// src/types/enums.ts
export enum EthereumTokens {
    USDT = 'usdt',
    USDC = 'usdc',
    BUSD = 'busd'
}

export enum Blockhains {
    Ethereum = 'ethereum',
    Aptos = 'aptos',
    Ton = 'ton'
}

export enum TonTokens {
    Ton = 'ton'
}

export enum AptosTokens {
    Apt = 'ton'
}

export enum EthereumNetworks {
    Polygon = 'polygon',
    Ethereum = 'ethereum',
    Bsc = 'bsc',
    Arbitrum = 'arbitrum'
}

export enum TestChains {
    Ethereum = 1,
    Polygon = 137,
    Bsc = 56,
    Arbitrum = 42161
}

export enum Chains {
    Ethereum = 5,
    Polygon = 80001,
    Bsc = 97,
    Arbitrum = 421613
}

export enum Allowance {
    Allowed = 'Allowed',
    NotAllowed = 'NotAllowed'
}